Add labels to activities table, delete unused view, make actitivities pack
