+++ app/javascript/packs/activities.js
@@ -1,0 +1,3 @@
+import { initTokenClickables } from "util.js";
+
+window.dodona.initTokenClickables = initTokenClickables;
