+++ app/views/activities/_series_activities_add_table.html.erb
