+++ app/views/activities/_series_activities_table.erb
@@ -1,6 +1,3 @@
-<% content_for :javascripts do %>
-  <%= javascript_pack_tag 'activities' %>
-<% end %>
 <% get_activity_path ||= method(:activity_path) %>
 <% if defined? user
      user = user || current_user
@@ -15,7 +12,6 @@
       <th class='status-icon'></th>
       <th class='type-icon'></th>
       <th><%= t "activities.index.activity_title" %></th>
-      <th><%= t "activities.index.labels" %></th>
       <% if @course.blank? || local_assigns[:series]&.progress_enabled || current_user&.course_admin?(@course) %>
         <th class='count hidden-xs'>
           <% if @course %>
@@ -51,7 +47,7 @@
 
         <td class='type-icon'>
           <% if activity.exercise? %>
-            <i class="<%= "mdi mdi-#{activity.programming_language&.icon} mdi-18" %>" title="<%= "#{t "activities.index.type.exercise_language", language: activity.programming_language&.name&.titleize}" %>"></i>
+            <i class="<%= "mdi mdi-#{activity.programming_language&.icon} mdi-18" %>" title="<%= "#{t "activities.index.type.exercise_language", language: activity.programming_language.name.titleize}" %>"></i>
           <% elsif activity.content_page? %>
             <i class="mdi mdi-book-open-variant mdi-18" title="<%= t "activities.index.type.content" %>"></i>
           <% end %>
@@ -65,21 +61,6 @@
           <% else %>
             <%= activity.name %>
           <% end %>
-        </td>
-
-        <td class="series-activities-table-labels">
-          <span class="<%= activity.programming_language ? "clickable-token" : "" %>" data-type="programming_languages" data-name="<%= activity.programming_language&.name %>"><%= activity.programming_language&.name || "" %></span>
-          <div class="text-muted">
-            <small>
-              <%# We can't just use activity.labels because a search by labels might restrict the loaded labels %>
-              <% labels = Label.includes(:activities).where(activities: {id: activity.id}).load %>
-              <% if labels.count > 5 %>
-                <span title="<%= labels.map(&:name).join(", ") %>"><%= raw labels.first(5).map {|label| "<span class='clickable-token' data-type='labels' data-name='#{label.name}'>#{label.name}</span>"}.join(", ") + ", ..." %></span>
-              <% else %>
-                <%= raw labels.map {|label| "<span class='clickable-token' data-type='labels' data-name='#{label.name}'>#{label.name}</span>"}.join(", ") %>
-              <% end %>
-            </small>
-          </div>
         </td>
 
         <% if @course.blank? || local_assigns[:series]&.progress_enabled || current_user&.course_admin?(@course) %>
@@ -141,8 +122,3 @@
 <% if activities.try(:total_pages) %>
   <center><%= page_navigation_links activities, true, "activities" %></center>
 <% end %>
-<script type="text/javascript">
-    $(function () {
-        dodona.initTokenClickables();
-    });
-</script>
