--- app/views/activities/_series_exercises_table.html.erb
@@ -1,71 +1,0 @@
-<table class="table activities-table table-resource">
-  <colgroup>
-    <col/>
-    <col/>
-    <col class="col-xs-5"/>
-    <col class="col-xs-5"/>
-    <col class="col-xs-1"/>
-  </colgroup>
-  <tbody>
-  <% activities.each do |activity| %>
-    <tr>
-      <td>
-        <% if local_assigns.fetch(:show_buttons_for_present, false) %>
-          <div class="drag-handle">
-            <i class="mdi mdi-reorder-horizontal mdi-18"></i>
-          </div>
-        <% end %>
-      </td>
-      <td class="status-icon">
-        <span class="pull-right">
-          <%= raw "<i class='mdi mdi-eye-off mdi-18' title='#{t("activities.index.private")}'></i>" if activity.access_private? %>
-          <%= raw "<i class='mdi mdi-delete mdi-18' title='#{t("activities.index.removed")}'></i>" if activity.removed? %>
-          <%= raw "<i class='mdi mdi-alert-circle mdi-18' title='#{t("activities.index.invalid")}'></i>" if activity.not_valid? %>
-        </span>
-      </td>
-      <td class="link">
-        <span class="ellipsis-overflow" title="<%= activity.name %>">
-          <% if local_assigns.fetch(:show_buttons_for_present, false) %>
-            <%= link_to activity.name, course_series_activity_path(series.course, series, activity), target: "_blank" %>
-          <% else %>
-            <%= link_to activity.name, activity_path(activity), target: "_blank" %>
-          <% end %>
-        </span>
-        <div class="text-muted ellipsis-overflow">
-          <small>
-            <span class="clickable-token" title="<%= activity.repository.name %>: <%= activity.path %>" data-type="repositories" data-name="<%= activity.repository.name %>"><%= activity.repository.name %>: <%= activity.path %></span>
-          </small>
-        </div>
-      </td>
-      <td class="series-activities-table-labels">
-        <span class="<%= activity.programming_language ? "clickable-token" : "" %>" data-type="programming_languages" data-name="<%= activity.programming_language&.name %>"><%= activity.programming_language&.name || "" %></span>
-        <div class="text-muted">
-          <small>
-            <%# We can't just use activity.labels because a search by labels might restrict the loaded labels %>
-            <% labels = Label.includes(:activities).where(activities: {id: activity.id}).load %>
-            <% if labels.count > 5 %>
-              <span title="<%= labels.map(&:name).join(", ") %>"><%= raw labels.first(5).map {|label| "<span class='clickable-token' data-type='labels' data-name='#{label.name}'>#{label.name}</span>"}.join(", ") + ", ..." %></span>
-            <% else %>
-              <%= raw labels.map {|label| "<span class='clickable-token' data-type='labels' data-name='#{label.name}'>#{label.name}</span>"}.join(", ") %>
-            <% end %>
-          </small>
-        </div>
-      </td>
-      <td class="actions">
-        <% if local_assigns.fetch(:show_buttons_for_present, false) %>
-          <a href="#" class="btn btn-icon remove-activity" title="<%= t "series.edit.remove_activity" %>" data-series_id="<%= series.id %>" data-scoped_url="<%= course_series_activity_path(series.course, series, activity)%>" data-activity_id="<%= activity.id %>" data-activity_name="<%= activity.name %>">
-            <i class="mdi mdi-delete mdi-18"></i>
-          </a>
-        <% else %>
-          <a href="#" class="btn btn-icon add-activity <%= "hidden" if series.activities.include? activity %>" title="<%= t "series.edit.add_activity" %>" data-series_id="<%= series.id %>" data-scoped_url="<%= course_series_activity_path(series.course, series, activity) %>" data-activity_id="<%= activity.id %>" data-activity_name="<%= activity.name %>" data-confirm="<%= t('.course_added_to_usable') unless activity.usable_by? @series.course %>">
-            <i class="mdi mdi-plus mdi-18"></i>
-          </a>
-        <% end %>
-      </td>
-    </tr>
-  <% end %>
-  </tbody>
-</table>
-<% if local_assigns.fetch(:paginate, true) %>
-  <center><%= page_navigation_links activities, true, "activities", {series_id: series.id}, "available" %></center>
-<% end %>
